task1/requirement.txt
python==3.12.6
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5